Moved Mixins to common package Added @Pseudo to MixinSpeakerPeripheral Reverted build.gradle changes to exclude CC repos and dependencies
