+++ fabric/build.gradle
@@ -24,16 +24,6 @@ loom {
     accessWidenerPath = project(":common").loom.accessWidenerPath
 }
 
-repositories {
-    maven {
-        url "https://cursemaven.com"
-        content {
-            includeGroup "curse.maven"
-        }
-    }
-    maven { url "https://maven.shedaniel.me/" }
-}
-
 dependencies {
     modImplementation("net.fabricmc:fabric-loader:${rootProject.fabric_loader_version}")
 
@@ -51,11 +41,6 @@ dependencies {
     modImplementation("net.fabricmc:fabric-language-kotlin:1.8.1+kotlin.1.7.0")
 
     modImplementation("com.terraformersmc:modmenu:1.16.23")
-
-    // CC Restitched
-    modImplementation("curse.maven:cc-restitched-462672:3622561")
-    modApi "me.shedaniel.cloth.api:cloth-api:1.5.50"
-    include "me.shedaniel.cloth.api:cloth-api:1.5.50"
 
     // Depend on the fabric API
     modImplementation("net.fabricmc.fabric-api:fabric-api:${rootProject.fabric_api_version}")
