+++ fabric/src/main/resources/valkyrienskies-fabric.mixins.json
@@ -5,8 +5,6 @@
   "plugin": "org.valkyrienskies.mod.fabric.mixin.ValkyrienFabricMixinConfigPlugin",
   "mixins": [
     "feature.explosions.ClipContextMixin",
-    "mod_compat.cc_restitched.MixinSpeakerPeripheral",
-    "mod_compat.cc_restitched.MixinWirelessNetwork",
     "world.level.block.FireMixin"
   ],
   "client": [
